Update plugin to latest versions of things
IJ platform 203
gradle 0.6.4
kotlin 1.4.0

This makes the tests fail. There is some issue w/ not being able to find
a class in the markdown plugin when using the 203 EAP builds. The plugin
works just run when task runIde is launched.

If gradle or kotlin is not upgraded (and only IJ platform is upgraded),
then the following issue occurs:
https://youtrack.jetbrains.com/issue/KT-37435

So there is some problem w/ the combination of these 3 changes that is
resulting in tests basically not the classes for the markdown plugin!
Need to wait until this issue in something not working out between
the gradle plugin for intellij, kotlin version, and platform classes.